+++ CHANGELOG.md
@@ -8,8 +8,12 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ### Added
 * thickness to `Boxes.draw_on`
-* add `round` to Boxes, Points, and Coords
-* add `contains` to Boxes
+* add `.round` to Boxes, Points, and Coords
+* add `Boxes.contains`
+* add `kwiamge.load_image_shape` 
+* add `MultiPolygon.fill` 
+* `kwimage.imwrite` now accepts the `backend` keyword, which can be `cv2`, `skimage`, or `gdal`. The `gdal` backend writes images as cloud-optimized-geotiffs.
+* add `kwimage.structs.Segmentation` which encapsulates Masks, Polygons, and MultiPolygons
 
 ### Fixed:
 * issue with `draw_clf_on_image` when truth is unknown